fix: remove circular refs - alpha.17
