+++ lib/src/dataservices/interfaces.ts
@@ -1,0 +1,14 @@
+export type HttpMethods = 'DELETE' | 'GET' | 'POST' | 'PUT';
+
+export interface RequestData {
+  method: HttpMethods;
+  url: string;
+  options: any;
+}
+
+/**
+ * A key/value map of parameters to be turned into an HTTP query string
+ * Same as HttpClient's HttpParamsOptions which is NOT exported at package level
+ * https://github.com/angular/angular/issues/22013
+ */
+export interface QueryParams { [name: string]: string | string[]; }
