+++ lib/src/entity-metadata/entity-definition.service.ts
@@ -2,7 +2,7 @@ import { Inject, Injectable, InjectionToken, Optional } from '@angular/core';
 
 import { createEntityDefinition, EntityDefinition } from './entity-definition';
 import { EntityMetadata, EntityMetadataMap, ENTITY_METADATA_TOKEN } from './entity-metadata';
-import { ENTITY_CACHE_NAME } from '../reducers';
+import { ENTITY_CACHE_NAME } from '../reducers/constants';
 
 export interface EntityDefinitions { [entityName: string]: EntityDefinition<any> }
