+++ lib/src/utils/index.ts
@@ -1,3 +1,3 @@
-export * from './utils';
 export * from './ngrx-entity-models'; // should be exported by @ngrx/entity
 export * from './pluralizer';
+export * from './utilities';
